refactor(ticketing): extract bulk/cascade into ticketing/bulk and collapse facade (resolves #1868)
